refactor(ldap+types): :truck: Move ldap types to types
